proposal for a minimal wraith with . that copes with unicode
Explanation:

# no single-element special case

The special case of treating a single element on a line as not a
function call leads to the need for more syntax down the line.

Introducing `&` was only required due to relying on this special case.

# . replaces \ and & again

If you use \ at the end of the line, the following is hard to follow:

define (display-excitement str reason r2)
  format "I'm SO EXCITED about ~a (~a, ~a)!!"
         string-upcase str \
         reason r2

ask your eyes: Is the reason a procedure call?

This becomes clear with the infix `.`:

define (display-excitement str reason r2)
  format "I'm SO EXCITED about ~a (~a, ~a)!!"
         string-upcase str
         . reason r2

This is the reason why the `.` exists in Wisp. Using the `.` avoids
the need for another syntax-character (the &) interspersed between
every argument.

# in-line indentation and unicode

Matching indentation must always be based on leading whitespace (only
spaces!), because indentation within a line cannot be calculated in
the general case. For example regional indicator symbols (flags) can
be one or two letters wide depending on whether the region is defined
on the system — and may change with system updates.

Therefore braces cannot have multi-line calls on the first line.

The following could not be parsed in the general case without complex backtracking:

define (greeter name)
  let [to-say
          format "Hey there ~a! :D"
                 . name
       show displayfn]
    displayln to-say

Therefore to-say must move on its own line so show can be matched up with it:

define (greeter name)
  let [
       to-say
          format "Hey there ~a! :D"
                 . name
       show displayfn]
    displayln to-say

One way to avoid this would be to state that this would be equivalent
to the code above:

define (greeter name)
  let [to-say
       format "Hey there ~a! :D"
                 . name
     show displayfn]
    displayln to-say

Upon hitting the less-indented show, it must treat format as nested
into to-say. I did not specify this, because I do not know whether it
can be implemented efficiently. It would mean that

Treating the first element as not-a-function-call would not work,
because a let with multiple arguments could become a valid structure,
even though it is currently an error:

(let ((a b c (values 1 2 3))) a)

An indentation-sensitive implementation should not prevent the future
use of not-yet-used structures, else it loses the advantage of
s-expressions.

That the let-expression above is currently an error means that the
path is open to giving it specific meaning in a future standard,
because there should not be existing competing implementations.